requirements.txt
numpy
opencv-python
scikit-image
scipy
pillow
pillow-heif
matplotlib
